dom/devicestorage/test/test_lastModificationFilter.html

<!--
  Any copyright is dedicated to the Public Domain.
  http://creativecommons.org/publicdomain/zero/1.0/
-->
<!DOCTYPE HTML>
<html> <!--
https://bugzilla.mozilla.org/show_bug.cgi?id=717103
-->
<head>
  <title>Test for the device storage API </title>
  <script type="text/javascript" src="/tests/SimpleTest/SimpleTest.js"></script>
  <script type="text/javascript" src="devicestorage_common.js"></script>

<link rel="stylesheet" type="text/css" href="/tests/SimpleTest/test.css" />
</head>
<body>
<a target="_blank" href="https://bugzilla.mozilla.org/show_bug.cgi?id=717103">Mozilla Bug 717103</a>
<p id="display"></p>
<div id="content" style="display: none">
  
</div>
<pre id="test">
<script class="testbody" type="text/javascript">

devicestorage_setup();

var oldFiles = ["a.png", "b.png", "c.png"];
var newFiles = ["d.png", "e.png", "f.png"];

var storage = navigator.getDeviceStorage('pictures');
var prefix = "devicestorage/" + randomFilename(12);
var callback;
var files;
var i;
var timestamp;

function verifyAndDelete(prefix, files, e) {
  if (e.target.result == null) {
    ok(files.length == 0, "when the enumeration is done, we shouldn't have any files in this array")
    dump("We still have length = " + files.length + "\n");
    dump(files + "\n");
    devicestorage_cleanup();
    return;
  }

  var filename = e.target.result.name;
  if (filename[0] == "/") {
    // We got /storageName/prefix/filename
    // Remove the storageName (this shows up on FirefoxOS)
    filename = filename.substring(1); // Remove leading slash
    var slashIndex = filename.indexOf("/");
    if (slashIndex >= 0) {
      filename = filename.substring(slashIndex + 1); // Remove storageName
    }
  }
  if (filename.startsWith(prefix)) {
    filename = filename.substring(prefix.length + 1); // Remove prefix
  }

  var index = files.indexOf(filename);
  ok(index > -1, "filename should be in the enumeration : " + e.target.result.name);
  if (index == -1)
    return;

  files.remove(index);

  // clean up
  var cleanup = storage.delete(e.target.result.name);
  cleanup.onsuccess = function(e) {}
}

function addSuccess(e) {
  i = i + 1;
  if (i == files.length) {
    callback();
    return;
  }
  addFile(files[i]);
}

function addError(e) {
  ok(false, "addError was called : " + e.target.error.name);
  devicestorage_cleanup();
}

function addFile(filename){
  var req = storage.addNamed(createRandomBlob('image/png'), prefix + '/' + files[i]);
  req.onsuccess = addSuccess;
  req.onerror = addError;
}

function afterNewFiles() {
  var cursor = storage.enumerate(prefix, {"since": timestamp});
  cursor.onsuccess = function(e) {
    verifyAndDelete(prefix, newFiles, e);
    if (e.target.result) {
      e.target.continue();
    }
  };
  cursor.onerror = function (e) {
    ok(false, "handleError was called : " + e.target.error.name);
    devicestorage_cleanup();
  };
}

function addNewFiles() {
  i = 0;
  files = newFiles;
  callback = afterNewFiles;
  addFile(files[0]);
}

function beforeNewFiles() {
  timestamp = new Date();
  setTimeout(addNewFiles, 1000);
}

function afterOldFiles() {
  setTimeout(beforeNewFiles, 1000);
}

function addOldFiles() {
  i = 0;
  files = oldFiles;
  callback = afterOldFiles;
  addFile(files[0]);
}

addOldFiles();

</script>
</pre>
</body>
</html>
